[jules] refactor: Use generic Skeleton component in FriendsScreen
Extracted the inline skeleton animation logic from FriendsScreen into a highly reusable `Skeleton` primitive in `mobile/components/ui/`. Refactored `FriendsScreen` to import and utilize the new generic component to produce a consistent and visually identical loading state while supporting light and dark themes out-of-the-box.
